Set tx.gasLimit to blockGasLimit on eth_estimateGas call
The gasLimit for an estimateGas call was set to the
defaultTransactionGasLimit of 90000 gas. This would cause
ethereumjs-vm runCall to throw the error 'base fee exceeds
gas limit'.

By setting the gasLimit to the block gas limit, this allows use to
correctly estimate gas for tx's where the basefee > gasLimit, such as
when deploying a contract.