Align block_helper.toJSON with hashed values
At OpenRelay, we're trying to use Ganache to test our services that
are written in Go. The RPC client that ships with Geth ignores the
hash provided by RPC and recalculates it instead. Due to some discrepancies
between the value used to calculate the hash and the values served over
RPC, the Go client is unable to compute the same hash as Ganache.

Two of the issues I ran into are in Ganache core. One was that the RPC
value for mixHash did not match the value used to calculate the checksum.
For reasons that are not clear even after reviewing the commit history,
rather than including the mixHash used to calculate the hash, block_helper
was including a string of repeating "10"s. I have changed this to reflect
the mixHash stored in the block header.

The other issue I ran into was extraData. If block.header.extraData was
an empty buffer, it would get encoded as "0x00", which decodes to a buffer
containing a single null byte. When a client uses that value to compute
the hash, they obviously get different results. This modifies
to.rpcDataHexString to pass empty buffers through as "0x" instead of
padding it with an extra 0 byte.

My mission of being able to compute the same hash with Ganache and  the Geth
RPC client has one more stop at ethereumjs-block, which defaults to having
an empty nonce instead of an appropriately sized buffer of null bytes.