Add tests for gas refunds
Also had to tweak the `add`/`transfer` gas estimation tests as on `transfer` we'll estimate `15000` too high due to the `transfer` call setting storage back to `0` (and we don't account for this `15000` refund in our estimate). To work around the issue I increased the amount initially added to the recipient to `10` so that transfering `5` out in the transfer test doesn't set it to `0`.