Pending transactions (#227)
eth_getTransactionByHash used to return null for pending transactions, but according to JSON RPC spec, if the transaction is pending only these fields should return null: `blockHash`, `blockNumber`, and `transactionIndex`. This PR fixes this bug so now we return the transaction object with the above mentioned fields as null for pending transactions. And for eth_getTransactionByReceipt on pending transactions, we return null.