Change transaction signature value encoding to QUANTITY
Even though the Ethereum JSON-RPC spec says that transaction signature
values R and S are DATA, but all clients encode them as QUANTITY. The
signature values are defined as numbers in the yellow paper, suggesting
QUANTITY as the correct encoding.

This change ensures ganache is compatible with Geth, Parity and any
future spec change.