breaking(legacyInstamine only): error messages are returned on the error's `data` field now...
Previously they were contain within a combination of the 'results: {[hash: string]: unknown}' and 'hashes: string[]' properties.

Also, only evm_mine and miner_start return an array for the 'data' field, as these are the only places where multiple transactions may be executed (this isn't _entirely_ true when a nonce is skipped and then the skipped nonce is executed, but this behavior wasn't supported in v2 anyway!).