Implement wallet passphrase and conditional encryption
User-supplied passphrase is now used to encrypt wallet. If no
passphrase is supplied and the `--lock` options is not specified,
initial accounts are not encrypted to improve startup speeds.

A keyfile record is now added for every account. Previously, this
wasn't done, making it impossible to actually unlock initial
accounts. This fixes #996.

Updates api with the above changes to the wallet.