Naive initial attempt to fix miner
Pre-"fix":
When a transaction is mined and it is found that the block's gas limit can
fit the block, the miner:
 - checks if there are more transactions from the same origin
   - if so, remove the origin's next best transaction and put it into the
      priced heap. *this process locks the transaction*
   - if not, just sort the priced heap
 - checks if the block has enough gas left to mine more
   - if so, continues in the while loop
   - if not, breaks out

The checks were performed in that order, which was an issue - if an origin
had more than one transaction, but the block is full, the origin's second
transaction that didn't fit into the block is now "locked". This causes
the transaction to be stuck in the pool forever.

Post-"fix":
In the same scenario described above, the order is rearranged:
 - if there are no more transactions from the origin, just sort the
    priced heap
 - if there is not enough gas left in the block to mine more, break out
 - if there are more transactions from the origin, remove the origin's
    next best transaction and put in into the priced heap, locking the
    transaction

While this fixes the indefinitely locked transaction issue, this prevents
the origin's next best transaction from being put into the priced heap.